+++ packages/scrabble-solver/src/components/LogoSplashScreen/LogoSplashScreen.module.scss
@@ -4,7 +4,7 @@ $loading-duration: 250ms;
 $loaded-duration: 100ms;
 $hiding-duration: 200ms;
 
-.splash {
+.logoSplashScreen {
   display: flex;
   align-items: center;
   justify-content: center;
