+++ packages/scrabble-solver/src/components/LogoSplashScreen/index.ts
@@ -1,0 +1,1 @@
+export { default } from './LogoSplashScreen';
