--- packages/scrabble-solver/src/components/Screen/index.ts
@@ -1,1 +1,0 @@
-export { default } from './Screen';
