--- packages/scrabble-solver/src/components/Splash/index.ts
@@ -1,1 +1,0 @@
-export { default } from './Splash';
