+++ packages/scrabble-solver/src/components/SplashScreen/SplashScreen.module.scss
@@ -1,4 +1,4 @@
-.screen {
+.splashScreen {
   position: fixed;
   top: 0;
   bottom: 0;
