+++ packages/scrabble-solver/src/components/SplashScreen/index.ts
@@ -1,0 +1,1 @@
+export { default } from './SplashScreen';
