+++ packages/scrabble-solver/src/components/index.ts
@@ -8,6 +8,7 @@ export { default as EmptyState } from './EmptyState';
 export { default as Key } from './Key';
 export { default as Loading } from './Loading';
 export { default as Logo } from './Logo';
+export { default as LogoSplashScreen } from './LogoSplashScreen';
 export { default as Modal } from './Modal';
 export { default as NavButtons } from './NavButtons';
 export { default as NotFound } from './NotFound';
@@ -17,10 +18,9 @@ export { default as Rack } from './Rack';
 export { default as Radio } from './Radio';
 export { default as Results } from './Results';
 export { default as ResultsInput } from './ResultsInput';
-export { default as Screen } from './Screen';
 export { default as Solver } from './Solver';
 export { default as SolverMobile } from './SolverMobile';
-export { default as Splash } from './Splash';
+export { default as SplashScreen } from './SplashScreen';
 export { default as SquareButton } from './SquareButton';
 export { default as SvgFontCss } from './SvgFontCss';
 export { default as SvgFontFix } from './SvgFontFix';
